refactor(schema): split from.ts into many files
