+++ packages/specs/schema/src/decorators/common/anyOf.spec.ts
@@ -1,6 +1,7 @@
 import {JsonEntityStore} from "../../domain/JsonEntityStore.js";
+import {number} from "../../fn/number.js";
+import {string} from "../../fn/string.js";
 import {AllOf, getJsonSchema, getSpec, In, OneOf, OperationPath, Path, Property, SpecTypes} from "../../index.js";
-import {number, string} from "../../utils/from.js";
 import {AnyOf} from "./anyOf.js";
 
 describe("@AnyOf", () => {
