+++ packages/specs/schema/src/decorators/generics/genericOf.ts
@@ -1,7 +1,7 @@
 import {isObject, Type} from "@tsed/core";
 
 import {JsonEntityStore} from "../../domain/JsonEntityStore.js";
-import {string} from "../../utils/from.js";
+import {string} from "../../fn/string.js";
 import {GenericValue} from "../../utils/generics.js";
 
 /**
