+++ packages/specs/schema/src/decorators/operations/returns.ts
@@ -18,9 +18,9 @@ import {JsonEntityStore} from "../../domain/JsonEntityStore.js";
 import {JsonParameter} from "../../domain/JsonParameter.js";
 import {JsonResponse} from "../../domain/JsonResponse.js";
 import {JsonSchema, JsonSchemaObject} from "../../domain/JsonSchema.js";
+import {string} from "../../fn/string.js";
 import {JsonHeader, JsonHeaders} from "../../interfaces/JsonOpenSpec.js";
 import {getStatusModel} from "../../utils/defineStatusModel.js";
-import {string} from "../../utils/from.js";
 import {GenericValue} from "../../utils/generics.js";
 import {isSuccessStatus} from "../../utils/isSuccessStatus.js";
 import {mapHeaders} from "../../utils/mapHeaders.js";
