+++ packages/specs/schema/src/fn/integer.ts
@@ -1,0 +1,12 @@
+import {from} from "./from.js";
+
+/**
+ * Declare a new integer model.
+ *
+ * See @@JsonSchema@@ to discover available methods.
+ *
+ * @function
+ */
+export function integer() {
+  return from(Number).integer();
+}
