+++ packages/specs/schema/src/fn/object.ts
@@ -1,0 +1,11 @@
+import {JsonSchema} from "../domain/JsonSchema.js";
+import {from} from "./from.js";
+
+/**
+ * Declare a new object model.
+ *
+ *  * See @@JsonSchema@@ to discover available methods.
+ */
+export function object(properties: {[key: string]: JsonSchema} = {}): JsonSchema {
+  return from(Object).properties(properties);
+}
