+++ packages/specs/schema/src/index.ts
@@ -138,6 +138,23 @@ export * from "./domain/JsonRequestBody.js";
 export * from "./domain/JsonResponse.js";
 export * from "./domain/JsonSchema.js";
 export * from "./domain/SpecTypes.js";
+export * from "./fn/allOf.js";
+export * from "./fn/any.js";
+export * from "./fn/anyOf.js";
+export * from "./fn/boolean.js";
+export * from "./fn/collection.js";
+export * from "./fn/date.js";
+export * from "./fn/email.js";
+export * from "./fn/enums.js";
+export * from "./fn/from.js";
+export * from "./fn/integer.js";
+export * from "./fn/lazyRef.js";
+export * from "./fn/number.js";
+export * from "./fn/object.js";
+export * from "./fn/oneOf.js";
+export * from "./fn/string.js";
+export * from "./fn/uri.js";
+export * from "./fn/url.js";
 export * from "./hooks/alterIgnore.js";
 export * from "./hooks/alterOneOf.js";
 export * from "./hooks/alterRequiredGroups.js";
@@ -151,7 +168,6 @@ export * from "./utils/buildPath.js";
 export * from "./utils/concatParameters.js";
 export * from "./utils/concatPath.js";
 export * from "./utils/defineStatusModel.js";
-export * from "./utils/from.js";
 export * from "./utils/generateSpec.js";
 export * from "./utils/generics.js";
 export * from "./utils/getComputedType.js";
